server/src/conversion_tool/requirements/base.txt
django==1.10
djangorestframework==3.5.3
Markdown==2.6.2
django-model-utils==2.3.1
django-cors-headers==1.1.0
bcrypt==2.0.0
psycopg2==2.6.1
pytz==2015.04
django-countries==4.0
xlsxwriter==0.9.3
xlrd==1.0.0
python-dateutil
numpy
django-rest-auth
djangorestframework-jwt
django-allauth
Pillow
matplotlib
pdfkit
numpy 
scipy 
ipython 
jupyter 
pandas 
sympy 
nose
geoip2